Add selection sort algorithm implementation
This commit adds the implementation of the selection sort algorithm in the file "2-selection_sort.c". It also includes a helper function "swap_ints" for swapping two integers in an array. The selection_sort function sorts an array of integers in ascending order using the selection sort algorithm. The array is printed after each swap.